pyproject.toml
[project]
name = "consensoor"
version = "0.1.0"
description = "Lightweight Python consensus layer client for local testing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "remerkleable>=0.1.28",
    "py_ecc>=7.0.0",
    "aiohttp>=3.9.0",
    "pyjwt>=2.8.0",
    "click>=8.1.0",
    "pyyaml>=6.0.0",
    "pycryptodome>=3.20.0",
    "libp2p>=0.2.0",
    "python-snappy>=0.7.0",
    "rlp>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
consensoor = "consensoor.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["consensoor*"]
